refactor exportable wrapper into model class definition
