+++ examples/nlp/language_modeling/megatron_gpt_export.py
@@ -38,83 +38,10 @@ from pytorch_lightning import Trainer
 from transformer_engine.common import recipe
 
 from nemo.collections.nlp.models.language_modeling.megatron_gpt_model import MegatronGPTModel
-from nemo.collections.nlp.modules.common.megatron.utils import get_ltor_masks_and_position_ids
 from nemo.collections.nlp.parts.nlp_overrides import NLPDDPStrategy, NLPSaveRestoreConnector
 from nemo.core.classes import Exportable
 from nemo.core.config import hydra_runner
-from nemo.core.neural_types import ChannelType, NeuralType
 from nemo.utils import logging
-
-
-class MegatronGPTExportableModel(torch.nn.Module, Exportable):
-    def __init__(self, model):
-        super().__init__()
-        self.model = model
-        self.dtype = None
-        if model.cfg['precision'] == 'bf16':
-            self.dtype = torch.bfloat16
-        elif int(model.cfg['precision']) == 32:
-            self.dtype = torch.float
-        elif int(model.cfg['precision']) == 16:
-            self.dtype = torch.float16
-        else:
-            raise ValueError(f"precision: {model.cfg['precision']} is not supported.")
-
-    def forward(self, tokens, position_ids, attention_mask):
-        assert tokens.shape == position_ids.shape
-        assert attention_mask.shape[2] == attention_mask.shape[3] == tokens.shape[1] == position_ids.shape[1]
-        with torch.autocast('cuda', dtype=self.dtype):
-            output_tensor = self.model.forward(
-                tokens=tokens.cuda(),
-                text_position_ids=position_ids.cuda(),
-                attention_mask=attention_mask.cuda(),
-                labels=None,
-            )
-
-        return output_tensor
-
-    def freeze(self):
-        for param in self.parameters():
-            param.requires_grad = False
-
-    def input_example(self, max_batch=1, max_dim=768, seq_len=6):
-        ids = [self.model.tokenizer.text_to_ids(text) for text in ["how is the weather on           Sunday"]]
-        id_tensors = [torch.unsqueeze(torch.LongTensor(id_list), dim=0) for id_list in ids]
-        masks_and_position_ids = [
-            get_ltor_masks_and_position_ids(id_tensor, self.model.tokenizer.eos_id, False, False, False)
-            for id_tensor in id_tensors
-        ]
-        for tokens, attn_mask_and_pos_ids in zip(id_tensors, masks_and_position_ids):
-            attn_mask, _, pos_ids = attn_mask_and_pos_ids
-            return tokens, pos_ids, attn_mask
-
-    def get_dynamic_axes(self):
-        dynamic_axes = {
-            'id_tensors': {0: "BS", 1: "sequence"},
-            'position_ids': {0: "BS", 1: "sequence"},
-            'attention_mask': {0: "BS", 2: "sequence", 3: "sequence"},
-        }
-        return dynamic_axes
-
-    @property
-    def input_types(self) -> Optional[Dict[str, NeuralType]]:
-        return {
-            "id_tensors": NeuralType(('B', 'T'), ChannelType()),
-            "position_ids": NeuralType(('B', 'T'), ChannelType()),
-            "attention_mask": NeuralType(('D', 'D', 'T', 'T'), ChannelType()),
-        }
-
-    @property
-    def output_types(self) -> Optional[Dict[str, NeuralType]]:
-        return {"log_probs": NeuralType(('B', 'T', 'D'), ChannelType())}
-
-    @property
-    def input_names(self) -> List[str]:
-        return ['id_tensors', 'position_ids', 'attention_mask']
-
-    @property
-    def output_names(self) -> List[str]:
-        return ['log_probs']
 
 
 @hydra_runner(config_path="conf", config_name="megatron_gpt_export")
@@ -169,40 +96,20 @@ def nemo_export(cfg):
         logging.error("Your NeMo model class ({}) is not Exportable.".format(model.__class__.__name__))
         sys.exit(1)
 
-    #
-    #  Add custom export parameters here
-    #
+    # Export
     check_trace = cfg.export_options.runtime_check
 
-    if (
-        cfg.export_options.cache_support
-        and hasattr(model, "encoder")
-        and hasattr(model.encoder, "export_cache_support")
-    ):
-        model.encoder.export_cache_support = True
-        logging.info("Caching support is enabled.")
-        model.encoder.setup_streaming_params()
-
-    fp8_recipe = recipe.DelayedScaling(margin=0, interval=1, fp8_format=recipe.Format.E4M3)
     try:
-        with torch.no_grad(), torch.inference_mode(), te.onnx_export(True), te.fp8_autocast(
-            enabled=True, fp8_recipe=fp8_recipe
-        ), warnings.catch_warnings():
-            warnings.filterwarnings(action='ignore', category=torch.jit.TracerWarning, module=r'.*')
-
-            model.to(device=cfg.export_options.device).freeze()
-            model.eval()
-            exportable_model = MegatronGPTExportableModel(model)
-
-            exportable_model.export(
-                onnx_out,
-                onnx_opset_version=cfg.export_options.onnx_opset,
-                do_constant_folding=cfg.export_options.do_constant_folding,
-                dynamic_axes=exportable_model.get_dynamic_axes(),
-                check_trace=check_trace,
-                check_tolerance=cfg.export_options.check_tolerance,
-                verbose=cfg.export_options.verbose,
-            )
+        model.to(device=cfg.export_options.device).freeze()
+        model.eval()
+        model.mgpt_wrapper().export(
+            onnx_out,
+            onnx_opset_version=cfg.export_options.onnx_opset,
+            do_constant_folding=cfg.export_options.do_constant_folding,
+            check_trace=check_trace,
+            check_tolerance=cfg.export_options.check_tolerance,
+            verbose=cfg.export_options.verbose,
+        )
     except Exception as e:
         logging.error(
             "Export failed. Please make sure your NeMo model class ({}) has working export() and that you have the latest NeMo package installed with [all] dependencies.".format(
